backend/requirements.txt
# Core web framework
flask>=2.0.0
flask-cors>=3.0.0

# Environment management
python-dotenv>=0.19.0

# AI/ML libraries
google-generativeai>=0.3.0
sentence-transformers>=2.2.0
scikit-learn>=1.0.0
numpy>=1.21.0

# Document processing
langchain>=0.0.300
langchain-community>=0.0.20
pypdf>=3.0.0

# Vector databases (optional)
pinecone>=3.0.0

# Utilities
requests>=2.25.0
python-multipart>=0.0.5

# Optional: Better embeddings (install separately if needed)
# openai>=1.0.0

# Development
gunicorn>=20.1.0